+++ my-school/src/Components/Dashboards/MainContainer.js
@@ -14,7 +14,7 @@ const MainContainer = () => {
   useEffect(() => {
     //the user will not be hard coded once we add dynamic routes and logins
     axios
-      .get("https://my-school-v1.herokuapp.com/api/users/1")
+      .get("https://my-school-v1.herokuapp.com/api/users/6")
       .then((res) => {
         setUser(res.data);
         getFamilyName(res.data.family_id);
